resource_site.go: only set repo attribute on read if there is a repo path
If a site resource does not have a repo defined, the Netlify API will still return BuildSettings in the read response payload.

Now we check for the RepoPath in the response. If this is not empty, it means the resource must have a repo defined, as it is a required attribute.

Fixes bug where a perpetual non-empty plan would be produced for sites without a repo defined.