Makefile: do not require $GOPATH
As of Go 1.11, we can no longer assume that environments with Go installed have $GOPATH set. Without this set, the Makefile would attempt to clone the website repo into /src, which will fail in most environments.

The website targets would also not update the local copy of the website repo to the latest version.

We now retrieve a fresh clone of the website repo whenever it is needed.

Co-Authored-By: kt <kt@katbyte.me>